refactor logging logic to library
