+++ HabitLogger/HabitLogger.slnx
@@ -1,3 +1,4 @@
 <Solution>
+  <Project Path="../LoggerEngine/LoggerEngine.csproj" />
   <Project Path="HabitLogger.csproj" />
 </Solution>
