+++ LoggerEngine/LoggerEngine.slnx
@@ -1,0 +1,3 @@
+<Solution>
+  <Project Path="LoggerEngine.csproj" />
+</Solution>
